Add more tests, move to an in-file format
