+++ Gemfile
@@ -1,4 +1,3 @@
 source 'https://rubygems.org'
 
 gem "minitest", "~> 4.0"
-gem "rake"
