+++ bigger_num_with_same_digits.rb
@@ -13,3 +13,11 @@ class BiggerNumWithSameDigits
     return num.join.to_i
   end
 end
+
+require 'test/unit'
+
+class BiggerNumWithSameDigitsTest < Test::Unit::TestCase
+  def test_algorithm
+    assert_equal 38627, BiggerNumWithSameDigits.perform(38276)
+  end
+end
