+++ dijkstra.rb
