+++ game_of_life.rb
