+++ knapsack.rb
