+++ permutations.rb
