--- test/test_helper.rb
@@ -1,3 +1,0 @@
-require "bundler/setup"
-require "test/unit"
-
